LayoutTests/editing/pasteboard/cleanup-on-move.html

<!DOCTYPE html>

<body>
<script src="../../resources/dump-as-markup.js"></script>
<script src="resources/select-and-drag.js"></script>
<div contenteditable="true">
<ol id="test">
<li id="one">one</li><li id="two">two</li><li id="three">three</li><li id="four">four</li>
</ol>
</div>

<script>
Markup.description('This test drags a selection of a couple of <li> and drops them after another <li> in the same list. It verifies that no empty <li> (actually with a <br> inside) are left after the moving operation.');

if (window.testRunner) {
    selectListItems("two", "three", 1);
    Markup.dump("test", "The original list looks like this. 'two' and 'three' are selected and are going to be dropped after 'four'");
    dragSelectionToTarget("two", "four");
    Markup.dump("test", "'two' and 'three' should appear as <li> after 'four' and no empty <li> (nor a <br> placeholder) should be there");
}
</script>
</body>
</html>
